chromeos_shutdown: unmount encrypted filesystems
When running with mount-encrypted, the filesystems that need to be
unmounted are different. Without this, the next mount of the encrypted
partition would have to go through journal replay. Also, move record
of umount failures to root of stateful partition since /var may be
separate now.

The platform_CleanShutdown test will be updated for the new failure
path separately.

BUG=None
TEST=x86-alex build, manual testing

Change-Id: If65bce72381705cd06d83f671cc10a8528b035e9
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21901
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>