Add in a sync call when we see a button press
This CL adds in a sync call to the button press handling code in powerm when it
is a power down event. This will force pending writes to the logs and other
activities to occur. This is intended to try to salvage as much information as
possible when the user has to use a long press on this button to cause a hard
shutdown.

There are some cases that this will miss. If we are straight up not receiving
signals related to the power button then this will never be trigged. Also if the
logging code is buffering within itself and not depending on the operating
system for write buffering, this won't flush those operations.

BUG=chromium-os:30431
TEST=Built package for x86-alex.
     Built image for x86-alex and confirmed that power button behaviour hasn't
     regressed.
     Pressed power button to shut down device and after restarting say messages
     about syncing in the old log.

Change-Id: I1bd6903dd7232400b83682c51a545205f91c4852
Reviewed-on: https://gerrit.chromium.org/gerrit/22107
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>