Fix 'legacy' bios support in postinst for arm platforms.
When an arm platform boots with "cros_legacy" that means it's booting
classic U-Boot, not syslinux. The postinstall handling was trying
to configure syslinux, and trying to configure it on /dev/sda12 and
not /dev/mmcblk0p12. This caused everything to just die.

This change modifies ConfigureInstall to use a detected device for the boot
partition, as well as turning off all non-U Boot boot handling for arm
platforms. It adds additional unit testing for ConfigureInstall to test
the device detection code.

TEST=New Unittests, NOT tested on hardware
BUG=chromium-os:30507

Change-Id: I9c177160c38a36b3b77101bf10f7e6f728fc5c5f
Reviewed-on: https://gerrit.chromium.org/gerrit/22062
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>