Make cryptohomed return whether a user mount is ephemeral
This CL extends the IsMountedForUser D-Bus method so that it also
reports whether a user mount is ephemeral or not. This is needed so
that when the browser crashes, it can check whether the current user
had logged in as ephemeral and sync back with reality.

No clients of the D-Bus interface are affected as the IsMountedForUser
method is not currently called by anyone in Chrome OS land.

BUG=chromium:127230
TEST=cryptohomed unit tests

Change-Id: I32b4f1b466ec98d88e760bdebf98937a5eb80c0b
Reviewed-on: https://gerrit.chromium.org/gerrit/22108
Commit-Ready: Bartosz Fabianowski <bartfab@chromium.org>
Reviewed-by: Bartosz Fabianowski <bartfab@chromium.org>
Tested-by: Bartosz Fabianowski <bartfab@chromium.org>