Remove the requirement for a power button from powerm
For some devkits they are not equiped with a traditional power button
device. Currently there is a check for whether or not the device has a power
button and if it doesn't this causes powerm to bail. This is really not what
should be happening, since lacking this device is not a critical failure. All
that should happen is that functionality that is related to it should not be
accessiable. Since all of the functionality for it is driven by messages from
the device, this is implicitly done via it not being present.

BUG=chromium-os:30730
TEST=Built package and image.
     Tested image in x86-alex for regressions.
     Tested that powerm stops dying all the time on daisy <TBD>.

Change-Id: Ia52716219a2354daa1824ba828657697308de688
Reviewed-on: https://gerrit.chromium.org/gerrit/22392
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>