Disable powerd monitor configuration on signal
As the first step in migrating the monitor configuration code from
powerd to Chrome, we need to enable a kill-switch for the existing
powerd support which can be called from Chrome.  The next step will then
be possible as Chrome can start using its support while disabling powerd
at runtime.  A third step will involve the removal of the support for
kUseNewMonitorConfigSignal and the permanent removal of the monitor
configuration code in powerd.
The one piece of functionality which still does need to live in powerd
is the backlight controller which sometimes needs to disable outputs.
Those events are now going to be passed to Chrome via
kSetScreenPowerSignal.  Once support for kUseNewMonitorConfigSignal can
be removed, the callsites of kSetScreenPowerSignal can also be
refactored as to use a less "round-about" path than is currently used in
this CL (deferred the refactoring to make it obvious that these events
are just implementing a glorified runtime switch).

BUG=chromium:126493
TEST=tested that manually sending the kUseNewMonitorConfigSignal with
dbus-send successfully disables the monitor reconfiguration logic
(tested on Lumpy).
CQ-DEPEND=I9aa451ba412f3cc9ce0cbf92d9abad24cded106a

Change-Id: Ibc2acc329528b381f53a2c894f0327401771c45f
Reviewed-on: https://gerrit.chromium.org/gerrit/22499
Reviewed-by: Jeff Disher <disher@chromium.org>
Tested-by: Jeff Disher <disher@chromium.org>
Commit-Ready: Jeff Disher <disher@chromium.org>