Implemented recovery from bad authorization data.
If the authorization data passed to OnLogin is not correct, private
objects cannot be used. In practice, the correctness of the
authorization data should be checked before loading a token but in the
event this happens the response is to do the following:
 1) Delete all objects, public and private.
 2) Initialize a new key hierarchy with the given auth data.

If initializing a key hierarchy fails either the first time it runs or
during this recovery step, the private object condition will be set to
avoid blocking incoming requests for private object operations.

BUG=chromium-os:30668
TEST=Ran all unit tests, including two new ones.
     Manually tested bad auth data cases using p11_replay.

Change-Id: I9ad20aa072dd08f9206f05ecf3c1559a842bde93