Added an authorization data hash.
Authorization data will only be sent to the TPM if it matches this hash.
This shields the TPM from bad authorization data which can cause it to
enter dictionary-attack-defense mode. Also, fixed a bug found during
testing: the private object mutex should not be signaled in the
ObjectPoolImpl::DeleteAll method.

Note: This CL is security-relevant. Previously a hash of the
authorization data was not stored.

BUG=chromium-os:30836
TEST=Ran all unit tests. Ran platform_Pkcs11ChangeAuthData autotest.

Change-Id: I7c65205e858f74f8a69e13cfe8ae00eb2de75bf2