Fix all the issues with MetricsStore
Currently the error handling in MetricsStore is a touch Byzantine and
does not actually cover all cases. Additionally the code in
MetricsStore is very CHECKy. These two issues are actually very
intertwined since the is_broken construct was introduced to allow
powerd to not hit CHECKs. After a few iterations on this CL it is
obvious I cannot resolve 29430 correctly without also resolving
31594. Thus I am just going to do it all at once and be done with
it. There is also some cleanup to make things more consistent or easier
to hack on.

BUG=chromium-os:29430
BUG=chromium-os:31594
TEST=Build power_manager and run unittests.
     Build image and run on device to confirm no regressions.

Change-Id: I98d6a36abf4c0dd1029f29accb3fba37168eeea2
Reviewed-on: https://gerrit.chromium.org/gerrit/24642
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>