shill: DeviceAdaptor: Return error for AddIPConfig
Return an error instead of just silently passing the
"/" path back when Device::AddIPConfig is called over
the DBus interface.  This can signal to Chrome that
this is not flimflam, and a different approach needs
to be used to setup static IP configurations.

BUG=chromium-os:31131
TEST=Manual: Try to perform static IP config in Chrome
(which does not know about shill yet).  Confirm the
following message in the Chrome log:
[21447:21459:340768216548:ERROR:object_proxy.cc(443)] Failed to call method: org.chromium.flimflam.Error.NotSupported: This function is deprecated in shill
[21447:21447:340768216869:WARNING:cros_network_functions.cc(1052)] Add IPConfig failed:
confirmed that Chrome does not crash as a result.

Change-Id: Id648b86fefe16e4ce6aef43a01ac82e56c9b505b
Reviewed-on: https://gerrit.chromium.org/gerrit/23150
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Greg Spencer <gspencer@chromium.org>