shill: WiFi: Scan frequently immediately after disconnect
At startup, and after disconnect, temporarily decrease
the scan interval in order to ensure that we've exhaustively
found all the networks we are looking for.  Scanning in
WiFi is not an exact science, and it's not possible to
guarantee that with one scan we have found all possible
networks.  Since we can scan relatively infrequently
while idle (to conserve power), be a more aggressive just
as we disconnect to give us the best chance to get back
online.

BUG=chromium-os:30984
TEST=New test network_WiFiRoaming/011RetryReconnectHidden
https://gerrit.chromium.org/gerrit/23231

Change-Id: If8fc83008aa5d99abe6af157b5bf47d3c8466632
Reviewed-on: https://gerrit.chromium.org/gerrit/23307
Commit-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>