shill: modify scope_logger_unittest to use an embedded instance, rather than the global lazy_instance.
If we use a lazy_instance in this test, then the state of the
object used in this test depend ons the order in which tests
run (because other tests can access and modify the lazy_instance).

Eliminate the execution order dependency by using a private instance
of the class.

BUG=None
TEST=unit tests

Change-Id: I3b76f20dfc4f900a7452084a001d4750b8fe2f03
Reviewed-on: https://gerrit.chromium.org/gerrit/24247
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>