Add command line option to postinst to specify target bios type.
This change refactors the bios type to be an enum on the InstallConfig
struct so it can be discovered or set in a centralized place. It also
adds a command line option to allow the bios type to be specified
instead of detected.

This is needed in some factory install situations where we boot with
one bios type, but then install another before rebooting.

BUG=chromium-os:31407
TEST=Unittests install/update to a CR48.

Change-Id: I607e9872fa245071a23ee3578fe82290f7493f79
Reviewed-on: https://gerrit.chromium.org/gerrit/24417
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>