Assume Chrome will handle monitor configuration
Historically, powerd handled all monitor configuration operations (due to
add/remove monitor events, user request, or power requirements).  This support
recently moved in Chrome but it was kept here, as well, to be selectively
disabled by the kUseNewMonitorConfigSignal DBus message when that change landed.
Now that it has landed a few days ago, support for this old code path can now be
removed as it is unused.
Note that this change merely implements the mechanical operation of removing the
trivially unreachable code and the hook for the DBus signal but a later change
will be made as a larger refactoring to address other unused logic which has
accumulated here and simplify the design around the small pieces of
functionality which have still remained.  This part of the change is being made
first as it is smaller, safer, and does remove the reference to
kUseNewMonitorConfigSignal, which can hopefully be removed in the future.

BUG=chromium:126493
TEST=Manually tested with an additional display and mode switching on Lumpy
(behaviour is the same before and after this change) and verified that the logs
indicate kUseNewMonitorConfigSignal is now being ignored by powerd (error
message will disappear once this signal is deprecated on the Chrome side).

Change-Id: Ibe1c9d58600c90551612a2318d9ff45e516d55f7
Reviewed-on: https://gerrit.chromium.org/gerrit/24626
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Tested-by: Jeff Disher <disher@chromium.org>
Commit-Ready: Jeff Disher <disher@chromium.org>