shill: wimax: When a device is enabled, obtain the current network set.
Previously, when a WiMAX device got enabled, it initiated a scan
hoping to get a new network list through a signal later. However, this
is racy because the device may have already obtained a network set so
no new NetworksChanged signal would be broadcast. This could happen if
shill crashes and restarts, for example.

So, instead, retrieve the current WiMaxManager.Device.Networks
property and then rely on the device to signal further changes to the
live network set.

BUG=chrome-os-partner:10014
TEST=unit tests

Change-Id: I373529c8e40d61670e0f3ae3b157b3b043487ecb
Reviewed-on: https://gerrit.chromium.org/gerrit/24716
Tested-by: Darin Petkov <petkov@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Darin Petkov <petkov@chromium.org>