Added handler for new kSetIsProjectingMethod
Added basic handling for the new kSetIsProjectingMethod DBus method.
The implementation now just takes the first argument as a boolean which
is true if projecting, false if not.  This then updates the value of the
flag in MonitorReconfigure which fires the projection_callback_ if the
value was changed.

BUG=chromium-os:31672
TEST=emerge chromeos-base/power_manager

Change-Id: Id1aac30005546cc660f3168ca22f0b80aec4bba1
Reviewed-on: https://gerrit.chromium.org/gerrit/25035
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Reviewed-by: Jeff Disher <disher@chromium.org>
Tested-by: Jeff Disher <disher@chromium.org>
Commit-Ready: Jeff Disher <disher@chromium.org>