Fix the initial value in the rolling average time estimate
This CL fixes a bug where the first value for the rolling average time
estimate is garbage for a session. This is occuring because though we
poll the power supply on intialization were not propegating that
information to the RollingAverage and the current entries that were
being passed up to chrome. While diagnosing this I cleaned up some of
the related code to make it easier to follow or being more logically
correct.

BUG=chromium-os:31633
TEST=Built package and ran unittests.
     Built image and ran, checking that there are no regressions.
     Added logging to the image to make sure that all the values for
     the averaged battery time estimate were correct.

Change-Id: I2f85cb94919e687e4dc2659dfae9c508a4cfd221
Reviewed-on: https://gerrit.chromium.org/gerrit/25141
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>