Fix update_engine scons building.
To build update_engine without an emerge, you need to run setup_dev_packages
to get the needed packages into the chroot, and then "scons" to do the
actual build.

However, setup_dev_packages has a broken dependancy. That dependancy is needed
to generate the test coverage reports, but not otherwise. Also, the SConstruct
file listed a "BASE_VER" for libchrome dependancies that was outdated and
overridden in the ebuild file.

This change removes the broken package depedancy, and updates the BASE_VER
so that it's relatively easy to build the update engine directly without doing
an emerge. That breaks the script to measure how much unittest coverage we
have, but that script is already broken.

Also, Gilad was right about static int's in one of Jay's previous CLs (and I
was wrong). How he was declaring some constants is causing problems for
commanline compiles, but doesn't for ebuild builds. So, this CL fixes that.

BUG=None
TEST=scons compelted.

Change-Id: I134f83528d8ad6fe3e504d98efb3d1030bfa3865
Reviewed-on: https://gerrit.chromium.org/gerrit/25252
Reviewed-by: Jay Srinivasan <jaysri@chromium.org>
Tested-by: Jay Srinivasan <jaysri@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Commit-Ready: Don Garrett <dgarrett@chromium.org>