power supply: Get AC status from battery status if needed
Sometimes the AC sysfs might show an invalid state.  When this happens,
use the battery status field to determine if the AC is plugged in.

Note that the battery status field is not used elsewhere by
power_supply.  It is only used as a backup option.

Also, if the battery is not found, the AC can be assumed to be plugged
in.

BUG=chromium-os:29898
TEST=unit tests pass

Change-Id: Idd85f23c46191b7707b26ef0d298f9e042865d99
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25257
Reviewed-by: Benson Leung <bleung@chromium.org>