Add -O1 to local build script by default.
This is the ordinary optimization level we use when locally building,
testing and doing coverage analysis on update_engine, so we might as
well etch it into the build script. It's a useful compromise between the
production/ebuild level (-O2 and other flags) and something that's still
debugable with (say) gdb.  It should eliminate some of the discrepancies
between the two builds, e.g. by making gcc optimize away const static
members declared and assigned in header files...

BUG=None
TEST=Builds project successfully

Change-Id: I0fa4a3fdc3fba5a556143061cbc56006676f754d
Reviewed-on: https://gerrit.chromium.org/gerrit/25389
Reviewed-by: Jay Srinivasan <jaysri@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>