Add ability to change window size to RollingAverage
Currently the max window size of a RollingAverage is set during its
initialization. This value is static throughout the life of the
object. To allow for tapering off from the rolling average to the raw
value when the battery is low, it is desirable for the window size to
shrink and then later expand as the battery life increases.

BUG=chromium-os:31834
TEST=Build power_manager package and run unit tests.
     Build image and confirmed that there were no regressions.

Change-Id: I0160a4c908e8ab38640b4c6d4bd54d31f6fbf703
Reviewed-on: https://gerrit.chromium.org/gerrit/25651
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Jon Kliegman <kliegs@chromium.org>