Use the kernel we are installing to detect default bios type.
We have been using current kernel to detect what type of bios to install for
in postinst, but for some factory installers, this can be wrong (we boot EFI,
install secure boot, reboot). So, this change uses the kernel we are planning
to boot into to decide how to run postinst.

BUG=chromium-os:31942 (should solve chromium-os:31407 as well)
TEST=image_to_live on a CR48.

Change-Id: I46ea0a397a1fc10aa363894b37e190f08d739038
Reviewed-on: https://gerrit.chromium.org/gerrit/25699
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>