Add tapering to the window size for battery time averages
This CL changes the code for updating the battery time averages to
taper from 10 samples when battery life > 60 minutes to 1 sample when
battery life <= 10 minutes. This is to allow for a gradual tapering of
the window size so that we don't go directly from a large window to
the raw value. We want to be using the raw value at the 10 minute
mark, since the delay in the rolling average is on that order
anywise.

This CL does not fix the other source of error that the user is
seeing. Currently we are reporting the time until the battery is
drained, when in reality the device will power down sooner than that
to ensure a clean shutdown. Future CLs will address this
discrepancy.

BUG=chromium-os:31834
TEST=Built package and ran unittests.
     Built image for device and confirmed no regressions.
     Ran down battery and saw that time values stayed accurate to
     within ~5 minutes at low levels.

Change-Id: I5393df761cf842063d53642beb377dd9d5a77e9d
Reviewed-on: https://gerrit.chromium.org/gerrit/25785
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>