laptop_mode_tools: add support for customization of the cpu governor
Add support for customization of the cpu governor configuration.
If a file /home/chronos/.governor_* is present, clone the read-only
configuration data to the /var/run/laptop-mode-tools directory and setup
the user config over top.  This is a temporary addition to enable testing
of the interactive governor.

BUG=chromium-os:17379
TEST=use crosh governor command to force interactive governor; reboot and verify the active governor is interactive; also tested other available governors

Change-Id: I37d6d5ab263210b980a395d70e7ead5015e5e96f
Reviewed-on: https://gerrit.chromium.org/gerrit/25795
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>
Commit-Ready: Sam Leffler <sleffler@chromium.org>