Trim off the low power threshold from the displayed time remaining
This trims off the threshold (180 seconds) from the battery remaining
time. This means that we are now passing up to the UI the amount of
time until the device is shut down, which is more useful thenn how
long until the battery is drained. We DO NOT adjust the percentage
value going up, since scaling this would be a bit of a nightmare and
we restrict the amount of error introduced by having the threshold be
less then a percent of the batter.

CQ-DEPEND=Ib012624ed629c8d9172e5c1790944a9ec5c67fbb
BUG=chromium-os:32160
TEST=Built package and ran unittests.
     Built image and confirmed no regressions.
     Ran device down and saw that the battery was reporting less then
     3 mins left in the UI before shutting down.

Change-Id: I64b86e87613f39e91ccb61e9343e4976d853626c
Reviewed-on: https://gerrit.chromium.org/gerrit/26328
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>