Enable low-memory notifier with compressed swap
We turned it off because it interfered with swap, but after
https://gerrit.chromium.org/gerrit/#change,26616
the notifier handles swap correctly.

The margin is left at its default of 50Mb.  The kernel
change weighs available swap space at a factor of 1/4
of RAM, so when RAM is consumed, the notifier will
trigger when 200Mb of (uncompressed) swap are left.

BUG=chromium-os:32322
TEST=manual

Change-Id: I978707bb5db186df33b2989548eb3f5b4830b71a
Reviewed-on: https://gerrit.chromium.org/gerrit/26709
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>