Improve audit-ability of "normal" mount state
This factors test-mode mount hacks out of chromeos_startup entirely,
and re-orders dev-mode mount changes to occur strictly after all
normal mount operations, enabling us to capture a clean log.

BUG=chromium-os:2285
TEST=run_remote_tests ... --use_emerged control.security\$
passes on ToT. This exercises the system in a bunch of ways,
including relying on +exec being enabled on various partitions.
Supplemented with ad hoc inspection of the mount properties
directly, and inspection of the log file contents, on
both a mod-for-test system, and a dev-image in dev-mode.

Change-Id: I43f8221cfbcf0fde9ce156e6346274c7923ef0fd
Reviewed-on: https://gerrit.chromium.org/gerrit/27233
Commit-Ready: Jim Hebert <jimhebert@chromium.org>
Reviewed-by: Jim Hebert <jimhebert@chromium.org>
Tested-by: Jim Hebert <jimhebert@chromium.org>