Fix typing issues in Input::EventHandler
In Input::EventHandler we currently store the return value of read in a unsigned
int. This is wrong, since read returns a ssize_t. When read fails it returns -1,
which when stored in a unsigned int becomes a very large positive value. Thus
the check that is supposed to catch this failure will never catch it. To
compound the issue we then promptly iterate over the memory received from
read. If the read returns -1 case this means we iterate over a huge chunk of
uninitialized memory, which is the likely source of the instability in powerm.

BUG=chromium-os:29535
TEST=Built package and ran unittests.
     Ran image with package and checked for regressions.
     After landing, I will watch the crash reports to see if stability improves
     <TBD>

Change-Id: I082b882e5254f2cf47be43ff3cd133595edf3a4d
Reviewed-on: https://gerrit.chromium.org/gerrit/27282
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Chris Wolfe <cwolfe@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>